# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Users Module API
  description: |
    REST API for the users-module. Manages user accounts, authentication,
    authorization, application tenancy, and audit logging.
  version: 0.1.0
  contact:
    email: zane@liquid-labs.com
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtained via any supported auth method (password, email-code,
        Google OIDC, Microsoft OIDC). Pass as `Authorization: Bearer <token>`.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          example: bad_request
        message:
          type: string
          description: Human-readable error description.
          example: "invalid JSON body"
...
    UserSummary:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        is_admin:
          type: boolean
        entity:
          $ref: "#/components/schemas/EntityInfo"
    EntityInfo:
      type: object
      properties:
        kind:
          type: string
          enum: [natural_person, corporation, service_account]
        display_name:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        legal_name:
          type: string
        jurisdiction:
          type: string
        label:
          type: string
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            default_app:
              $ref: "#/components/schemas/AppSummary"
...
    AdminCreateUserRequest:
      type: object
      required: [email, given_name, family_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        is_admin:
          type: boolean
          default: false
    AdminUpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        given_name:
          type: string
        family_name:
          type: string
        is_admin:
          type: boolean
...
    AppSummary:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
...
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
        total:
          type: integer
...
  parameters:
    UserUUID:
      name: user_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Caller lacks permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/users:
    get:
      summary: Search / list users (admin)
      tags: [Users]
      operationId: listUsers
      parameters:
        - name: q
          in: query
          description: Search query (matches email, name).
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated user list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create a user (admin)
      tags: [Users]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/users/{user_uuid}:
    get:
      summary: Get a user by UUID (admin)
      tags: [Users]
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a user (admin)
      tags: [Users]
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Archive a user (admin, soft-delete)
      tags: [Users]
      operationId: archiveUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      responses:
        "204":
          description: User archived.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...